dedup createTask
